requirements.txt
pygame-ce>=2.5.6
pyyaml>=6.0
